paint_api

Module display_list

Defines data structures which are consumed by Paint.

Structs

AxesScrollSensitivity

The ScrollType of particular node in the vertical and horizontal axes.

PaintDisplayListInfo

A data structure which stores Paint-side information about display lists sent to Paint.

ReferenceFrameNodeInfo

ScrollTree

A tree of spatial nodes, which mirrors the spatial nodes in the WebRender display list, except these are used for scrolling in Paint so that new offsets can be sent to WebRender.

ScrollTreeNode

A node in a tree of scroll nodes. This may either be a scrollable node which responds to scroll events or a non-scrollable one.

ScrollTreeNodeTransformationCache

A cached of transforms of a particular ScrollTree node in both directions: mapping from node-relative points to root-relative points and vice-versa.

ScrollType

A scroll type, describing whether what kind of action originated this scroll request. This is a bitflag as it is also used to track what kinds of ScrollTypes scroll nodes are sensitive to.

ScrollableNodeInfo

Data stored for nodes in the ScrollTree that actually scroll, as opposed to reference frames and sticky nodes which do not.

StickyNodeInfo

Enums

SpatialTreeNodeInfo

TouchAction

A simplified representation of the CSS touch-action property, used by the compositor to decide how a touch gesture may scroll a given node.

Type Aliases

AdjacencyListForPrint 🔒

In order to pretty print the ScrollTree structure, we are converting the node list inside the tree to be a adjacency list. The adjacency list then is used for the ScrollTree::debug_print_traversal of the tree.