paint_api::display_list

Struct PaintDisplayListInfo

pub struct PaintDisplayListInfo {
    pub pipeline_id: PipelineId,
    pub viewport_details: ViewportDetails,
    pub content_size: LayoutSize,
    pub epoch: Epoch,
    pub scroll_tree: ScrollTree,
    pub root_reference_frame_id: ScrollTreeNodeId,
    pub root_scroll_node_id: ScrollTreeNodeId,
    pub is_paintable: bool,
    pub is_contentful: bool,
    pub first_reflow: bool,
    pub caret_property_binding: Option<(PropertyBindingKey<ColorF>, ColorF)>,
}

A data structure which stores Paint-side information about display lists sent to Paint.

Fields
pipeline_id: PipelineId

The WebRender PipelineId of this display list.

viewport_details: ViewportDetails

The ViewportDetails that describe the viewport in the script/layout thread at the time this display list was created.

content_size: LayoutSize

The size of this display list’s content.

epoch: Epoch

The epoch of the display list.

scroll_tree: ScrollTree

A ScrollTree used by Paint to scroll the contents of the display list.

root_reference_frame_id: ScrollTreeNodeId

The ScrollTreeNodeId of the root reference frame of this info’s scroll tree.

root_scroll_node_id: ScrollTreeNodeId

The ScrollTreeNodeId of the topmost scrolling frame of this info’s scroll tree.

is_paintable: bool

From https://www.w3.org/TR/paint-timing/#paintable: Whether the display list contains paintable items.

is_contentful: bool

From https://www.w3.org/TR/paint-timing/#contentful: Contentful paint i.e. whether the display list contains items of type text, image, non-white canvas or SVG). Used by metrics.

first_reflow: bool

Whether the first layout or a subsequent (incremental) layout triggered this display list creation.

caret_property_binding: Option<(PropertyBindingKey<ColorF>, ColorF)>

If this display list contains a blinking caret, this value will be filled with its animation key and original color value so that the painter can animate the caret.

Implementations

impl PaintDisplayListInfo

pub fn new( viewport_details: ViewportDetails, content_size: LayoutSize, pipeline_id: PipelineId, epoch: Epoch, viewport_scroll_sensitivity: AxesScrollSensitivity, first_reflow: bool, ) -> Self

Create a new PaintDisplayListInfo with the root reference frame and scroll frame already added to the scroll tree.

pub fn external_scroll_id_for_scroll_tree_node( &self, id: ScrollTreeNodeId, ) -> ExternalScrollId