paint_api

Enum PaintMessage

pub enum PaintMessage {

    ChangeRunningAnimationsState(WebViewId, PipelineId, AnimationState),
    SetFrameTreeForWebView(WebViewId, SendableFrameTree),
    SetThrottled(WebViewId, PipelineId, bool),
    NewWebRenderFrameReady(PainterId, DocumentId, bool),
    PipelineExited(WebViewId, PipelineId, PipelineExitSource),
    SendInitialTransaction(WebViewId, PipelineId),
    ScrollNodeByDelta(WebViewId, PipelineId, LayoutVector2D, ExternalScrollId),
    ScrollViewportByDelta(WebViewId, LayoutVector2D),
    UpdateEpoch {
        webview_id: WebViewId,
        pipeline_id: PipelineId,
        epoch: Epoch,
    },
    SendDisplayList {
        webview_id: WebViewId,
        display_list_descriptor: BuiltDisplayListDescriptor,
        display_list_info_receiver: GenericReceiver<PaintDisplayListInfo>,
        display_list_data_receiver: GenericReceiver<SerializableDisplayListPayload>,
    },
    GenerateFrame(Vec<PainterId>),
    GenerateImageKey(WebViewId, GenericSender<ImageKey>),
    GenerateImageKeysForPipeline(WebViewId, PipelineId),
    UpdateImages(PainterId, SmallVec<[ImageUpdate; 1]>),
    DelayNewFrameForCanvas(WebViewId, PipelineId, Epoch, Vec<ImageKey>),
    GenerateFontKeys(usize, usize, GenericSender<(Vec<FontKey>, Vec<FontInstanceKey>)>, PainterId),
    AddFont(PainterId, FontKey, Arc<GenericSharedMemory>, u32),
    AddSystemFont(PainterId, FontKey, NativeFontHandle),
    AddFontInstance(PainterId, FontInstanceKey, FontKey, f32, FontInstanceFlags, Vec<FontVariation>),
    RemoveFonts(PainterId, Vec<FontKey>, Vec<FontInstanceKey>),
    CollectMemoryReport(ReportsChan),
    Viewport(WebViewId, ViewportDescription),
    ScreenshotReadinessReponse(WebViewId, FxHashMap<PipelineId, Epoch>),
    SendLCPCandidate(LCPCandidate, WebViewId, PipelineId, Epoch),

}

Messages from (or via) the constellation thread to Paint.

Variants

ChangeRunningAnimationsState(WebViewId, PipelineId, AnimationState)

Alerts Paint that the given pipeline has changed whether it is running animations.

SetFrameTreeForWebView(WebViewId, SendableFrameTree)

Updates the frame tree for the given webview.

SetThrottled(WebViewId, PipelineId, bool)

Set whether to use less resources by stopping animations.

NewWebRenderFrameReady(PainterId, DocumentId, bool)

WebRender has produced a new frame. This message informs Paint that the frame is ready. It contains a bool to indicate if it needs to composite, the DocumentId of the new frame and the PainterId of the associated painter.

PipelineExited(WebViewId, PipelineId, PipelineExitSource)

Script or the Constellation is notifying the renderer that a Pipeline has finished shutting down. The renderer will not discard the Pipeline until both report that they have fully shut it down, to avoid recreating it due to any subsequent messages.

SendInitialTransaction(WebViewId, PipelineId)

Inform WebRender of the existence of this pipeline.

ScrollNodeByDelta(WebViewId, PipelineId, LayoutVector2D, ExternalScrollId)

Scroll the given node (ExternalScrollId) by the provided delta. This will only adjust the node’s scroll position and will not do panning in the pinch zoom viewport.

ScrollViewportByDelta(WebViewId, LayoutVector2D)

Scroll the WebView’s viewport by the given delta. This will also do panning in the pinch zoom viewport if possible and the remaining delta will be used to scroll the root layer.

UpdateEpoch

Update the rendering epoch of the given Pipeline.

Fields

webview_id: WebViewId

The WebViewId that this display list belongs to.

pipeline_id: PipelineId

The PipelineId of the Pipeline to update.

epoch: Epoch

The new Epoch value.

SendDisplayList

Inform WebRender of a new display list for the given pipeline.

Fields

webview_id: WebViewId

The WebViewId that this display list belongs to.

display_list_descriptor: BuiltDisplayListDescriptor

A descriptor of this display list used to construct this display list from raw data.

display_list_info_receiver: GenericReceiver<PaintDisplayListInfo>

A GenericReceiver used to send the PaintDisplayListInfo.

display_list_data_receiver: GenericReceiver<SerializableDisplayListPayload>

A GenericReceiver used to send the serialized version of `DisplayListPayload.

GenerateFrame(Vec<PainterId>)

Ask the renderer to generate a frame for the current set of display lists from the given PainterIds that have been sent to the renderer.

GenerateImageKey(WebViewId, GenericSender<ImageKey>)

Create a new image key. The result will be returned via the provided channel sender.

GenerateImageKeysForPipeline(WebViewId, PipelineId)

The same as the above but it will be forwarded to the pipeline instead of send via a channel.

UpdateImages(PainterId, SmallVec<[ImageUpdate; 1]>)

Perform a resource update operation.

DelayNewFrameForCanvas(WebViewId, PipelineId, Epoch, Vec<ImageKey>)

Pause all pipeline display list processing for the given pipeline until the following image updates have been received. This is used to ensure that canvas elements have had a chance to update their rendering and send the image update to the renderer before their associated display list is actually displayed.

GenerateFontKeys(usize, usize, GenericSender<(Vec<FontKey>, Vec<FontInstanceKey>)>, PainterId)

Generate a new batch of font keys which can be used to allocate keys asynchronously.

AddFont(PainterId, FontKey, Arc<GenericSharedMemory>, u32)

Add a font with the given data and font key.

AddSystemFont(PainterId, FontKey, NativeFontHandle)

Add a system font with the given font key and handle.

AddFontInstance(PainterId, FontInstanceKey, FontKey, f32, FontInstanceFlags, Vec<FontVariation>)

Add an instance of a font with the given instance key.

RemoveFonts(PainterId, Vec<FontKey>, Vec<FontInstanceKey>)

Remove the given font resources from our WebRender instance.

CollectMemoryReport(ReportsChan)

Measure the current memory usage associated with Paint. The report must be sent on the provided channel once it’s complete.

Viewport(WebViewId, ViewportDescription)

A top-level frame has parsed a viewport metatag and is sending the new constraints.

ScreenshotReadinessReponse(WebViewId, FxHashMap<PipelineId, Epoch>)

Let Paint know that the given WebView is ready to have a screenshot taken after the given pipeline’s epochs have been rendered.

SendLCPCandidate(LCPCandidate, WebViewId, PipelineId, Epoch)

The candidate of largest-contentful-paint