paint_api::viewport_description

Struct ViewportDescription

pub struct ViewportDescription {
    pub initial_scale: Scale<f32, CSSPixel, DeviceIndependentPixel>,
    pub minimum_scale: Scale<f32, CSSPixel, DeviceIndependentPixel>,
    pub maximum_scale: Scale<f32, CSSPixel, DeviceIndependentPixel>,
    pub user_scalable: UserScalable,
}

A set of viewport descriptors:

https://www.w3.org/TR/css-viewport-1/#viewport-meta

Fields
initial_scale: Scale<f32, CSSPixel, DeviceIndependentPixel>

https://developer.mozilla.org/en-US/docs/Web/HTML/Viewport_meta_tag#initial-scale the zoom level when the page is first loaded

minimum_scale: Scale<f32, CSSPixel, DeviceIndependentPixel>

The minimum page zoom that is allowed on this viewport’s page. https://developer.mozilla.org/en-US/docs/Web/HTML/Viewport_meta_tag#minimum_scale

maximum_scale: Scale<f32, CSSPixel, DeviceIndependentPixel>

The maximum page zoom that is allowed on this viewport’s page. https://developer.mozilla.org/en-US/docs/Web/HTML/Viewport_meta_tag#maximum_scale

user_scalable: UserScalable

https://developer.mozilla.org/en-US/docs/Web/HTML/Viewport_meta_tag#user_scalable whether zoom in and zoom out actions are allowed on the page

Implementations

impl ViewportDescription

fn process_viewport_key_value_pairs( pairs: HashMap<String, String>, ) -> ViewportDescription

Iterates over the key-value pairs generated from meta tag and returns a ViewportDescription

fn parse_viewport_value_as_zoom( value: &str, ) -> Option<Scale<f32, CSSPixel, DeviceIndependentPixel>>

Parses a viewport zoom value.

pub fn clamp_zoom(&self, zoom: f32) -> f32

Constrains a zoom value within the allowed scale range